Updated UserDAO, implemented login and registration actors and formatted the code

+++ food-ministry-parent/food-ministry-actors/src/main/java/org/food/ministry/actors/factory/messages/UniqueIDResultMessage.java
@@ -3,7 +3,8 @@ package org.food.ministry.actors.factory.messages;
 import org.food.ministry.actors.messages.IMessage;
 
 /**
- * The message for delivering a before requested ID
+ * The message for delivering a requested ID
+ * 
  * @author Maximilian Briglmeier
  * @since 16.02.2019
  *
@@ -13,21 +14,24 @@ public class UniqueIDResultMessage implements IMessage {
     /**
      * The ID
      */
-    private int nextId;
-    
+    private long nextId;
+
     /**
      * Constructor setting the ID, which was requested before
-     * @param nextId the ID
+     * 
+     * @param nextId
+     *            the ID
      */
-    public UniqueIDResultMessage(int nextId) {
+    public UniqueIDResultMessage(long nextId) {
         this.nextId = nextId;
     }
 
     /**
      * Gets the ID, which was requested before
+     * 
      * @return The ID
      */
-    public int getNextId() {
+    public long getNextId() {
         return nextId;
     }
 }
