+++ food-ministry-parent/food-ministry-actors/src/main/java/org/food/ministry/actors/messages/AResultMessage.java
@@ -2,6 +2,7 @@ package org.food.ministry.actors.messages;
 
 /**
  * An abstract implementation of {@link IMessage} for results
+ * 
  * @author Maximilian Briglmeier
  * @since 16.02.2019
  *
@@ -11,11 +12,11 @@ public class AResultMessage implements IMessage {
     /**
      * The ID of this message
      */
-    private int id;
+    private long id;
     /**
      * The ID of the requesting message
      */
-    private int originId;
+    private long originId;
     /**
      * Determines if the requesting message was successfully processed
      */
@@ -24,15 +25,20 @@ public class AResultMessage implements IMessage {
      * Error message in case something went wrong
      */
     private String errorMessage;
-    
+
     /**
      * Constructor initializing the essential member variables
-     * @param id The ID of the message
-     * @param originId The ID of the requesting message 
-     * @param successful Determines if the requesting message was successfully processed
-     * @param errorMessage Error message in case something went wrong
+     * 
+     * @param id
+     *            The ID of the message
+     * @param originId
+     *            The ID of the requesting message
+     * @param successful
+     *            Determines if the requesting message was successfully processed
+     * @param errorMessage
+     *            Error message in case something went wrong
      */
-    public AResultMessage(int id, int originId, boolean successful, String errorMessage) {
+    public AResultMessage(long id, long originId, boolean successful, String errorMessage) {
         this.id = id;
         this.originId = originId;
         this.successful = successful;
@@ -41,23 +47,27 @@ public class AResultMessage implements IMessage {
 
     /**
      * Gets the ID of this message
+     * 
      * @return The ID of this message
      */
-    public int getId() {
+    public long getId() {
         return id;
     }
 
     /**
-     * Gets the ID of the requesting message 
-     * @return The ID of the requesting message 
+     * Gets the ID of the requesting message
+     * 
+     * @return The ID of the requesting message
      */
-    public int getOriginId() {
+    public long getOriginId() {
         return originId;
     }
-    
+
     /**
      * Determines if the requesting message was successfully processed
-     * @return true, if the requesting message was successfully processed; else false
+     * 
+     * @return true, if the requesting message was successfully processed; else
+     *         false
      */
     public boolean isSuccessful() {
         return successful;
@@ -65,6 +75,7 @@ public class AResultMessage implements IMessage {
 
     /**
      * Gets the error message
+     * 
      * @return The error message
      */
     public String getErrorMessage() {
