+++ food-ministry-parent/food-ministry-actors/src/main/java/org/food/ministry/actors/user/messages/AUserInfoMessage.java
@@ -3,7 +3,9 @@ package org.food.ministry.actors.user.messages;
 import org.food.ministry.actors.messages.IMessage;
 
 /**
- * An abstract implementation of {@link IMessage} for requesting user information
+ * An abstract implementation of {@link IMessage} for requesting user
+ * information
+ * 
  * @author Maximilian Briglmeier
  * @since 16.02.2019
  *
@@ -13,7 +15,7 @@ public class AUserInfoMessage implements IMessage {
     /**
      * The ID of this message
      */
-    private int id;
+    private long id;
     /**
      * The name of the user requesting information
      */
@@ -26,15 +28,20 @@ public class AUserInfoMessage implements IMessage {
      * The password of the user needed for user actions
      */
     private String password;
-    
+
     /**
      * Constructor initializing the message with essential member variables
-     * @param id The ID of this message
-     * @param username The name of the user requesting information
-     * @param emailAddress The email address of the user requesting information
-     * @param password The password of the user needed for user actions
+     * 
+     * @param id
+     *            The ID of this message
+     * @param username
+     *            The name of the user requesting information
+     * @param emailAddress
+     *            The email address of the user requesting information
+     * @param password
+     *            The password of the user needed for user actions
      */
-    public AUserInfoMessage(int id, String username, String emailAddress, String password) {
+    public AUserInfoMessage(long id, String username, String emailAddress, String password) {
         this.id = id;
         this.username = username;
         this.emailAddress = emailAddress;
@@ -43,14 +50,16 @@ public class AUserInfoMessage implements IMessage {
 
     /**
      * Gets the ID of this message
+     * 
      * @return
      */
-    public int getId() {
+    public long getId() {
         return id;
     }
 
     /**
      * Gets the name of the user requesting information
+     * 
      * @return The name of the user requesting information
      */
     public String getUsername() {
@@ -58,8 +67,9 @@ public class AUserInfoMessage implements IMessage {
     }
 
     /**
-     * Gets the email address of the user requesting information 
-     * @return The email address of the user requesting information 
+     * Gets the email address of the user requesting information
+     * 
+     * @return The email address of the user requesting information
      */
     public String getEmailAddress() {
         return emailAddress;
@@ -67,6 +77,7 @@ public class AUserInfoMessage implements IMessage {
 
     /**
      * Gets the password of the user needed for user actions
+     * 
      * @return The password of the user needed for user actions
      */
     public String getPassword() {
