+++ food-ministry-parent/food-ministry-actors/src/main/java/org/food/ministry/actors/user/messages/DelegateMessage.java
@@ -3,45 +3,51 @@ package org.food.ministry.actors.user.messages;
 import org.food.ministry.actors.messages.IMessage;
 
 /**
- * A message for delegating a request to another actor 
+ * A message for delegating a request to another actor
+ * 
  * @author Maximilian Briglmeier
  * @since 16.02.2019
  *
  */
 public class DelegateMessage implements IMessage {
-    
+
     /**
      * The ID of this message
      */
-    private int id;
+    private long id;
     /**
      * The ID of the original message
      */
-    private int originId;
-    
+    private long originId;
+
     /**
      * Constructor initializing the essential member variables
-     * @param id The ID of this message
-     * @param originId The ID of the original message
+     * 
+     * @param id
+     *            The ID of this message
+     * @param originId
+     *            The ID of the original message
      */
-    public DelegateMessage(int id, int originId) {
+    public DelegateMessage(long id, long originId) {
         this.id = id;
         this.originId = originId;
     }
 
     /**
      * Gets the ID of this message
+     * 
      * @return The ID of this message
      */
-    public int getId() {
+    public long getId() {
         return id;
     }
 
     /**
      * Gets the ID of the original message
+     * 
      * @return The ID of the original message
      */
-    public int getOriginId() {
+    public long getOriginId() {
         return originId;
     }
 }
