+++ food-ministry-parent/food-ministry-actors/src/main/java/org/food/ministry/actors/user/messages/LoginMessage.java
@@ -1,7 +1,9 @@
 package org.food.ministry.actors.user.messages;
 
 /**
- * A specific kind of {@link AUserInfoMessage} for transmitting the information for a login attempt.
+ * A specific kind of {@link AUserInfoMessage} for transmitting the information
+ * for a login attempt.
+ * 
  * @author Maximilian Briglmeier
  * @since 16.02.2019
  *
@@ -11,8 +13,8 @@ public class LoginMessage extends AUserInfoMessage {
     /**
      * @see AUserInfoMessage#AUserInfoMessage(int, String, String, String)
      */
-    public LoginMessage(int id, String username, String emailAddress, String password) {
+    public LoginMessage(long id, String username, String emailAddress, String password) {
         super(id, username, emailAddress, password);
     }
-    
+
 }
